change assignment1 to class
