--- cdf/utils/README.md
@@ -1,11 +1,0 @@
-# Utils 
-
-The utils folder in this repository contains various utility functions and scripts that support the main functionalities of the project. These utilities are designed to be reusable.
-
-**acs_pull.py** - This script focuses on extracting predefined sociodemographic data from the ACS for the Chicago.
-
-**acs_agg.py** - This script contains functions that aggregate census data based on library assignments
-
-**plotting.py** - This script contains functions for plotting folium maps
-
-**utils.py** - This script contains funcitons for reading data and components essential to the development of methodologies.
